Cleaned up and componentized the Carousel component.
