+++ .eslintrc
@@ -26,6 +26,7 @@
     "jsx-a11y/anchor-is-valid": false,
     "jsx-a11y/label-has-associated-control": false,
     "jsx-a11y/click-events-have-key-events": false,
-    "jsx-a11y/no-static-element-interactions": false
+    "jsx-a11y/no-static-element-interactions": false,
+    "import/prefer-default-export": false
   }
 }
